extract helper fns
